actualizacion del 10/05/2021

+++ src/app/usuario/page/page.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
 import { ServiceService } from "src/app/Service/service.service";
 import { Usuario } from '../../Modelo/usuario';
+import swal from 'sweetalert2';
 
 @Component({
   selector: "app-getPageUsuario",
@@ -10,28 +11,118 @@ import { Usuario } from '../../Modelo/usuario';
 })
 export class getPageUsuario implements OnInit{
 
-    usuario!:Usuario[];
+    usuario!:Array<Usuario>;
+    totalPages!: Array<number>;
+
+    page = 0;
+    size = 4;
+    order ='id';
+    asc = true;
+
+    isFirst = false;
+    isLast = false;
+
     constructor(private service:ServiceService, private router:Router){}
 
     ngOnInit(){
-      this.service.getPageUsuario().subscribe(data=>{this.usuario=data;})
+      this.cargarUsuarios();
 
     }
 
+    cargarUsuarios(){
+      this.service.getPageUsuario(this.page, this.size, this.order, this.asc).subscribe
+        (data=>{
+
+          console.log(data);
+
+          this.usuario = data.content;
+          this.isFirst = data.first;
+          this.isLast = data.last;
+          this.totalPages = new Array(data.totalPages);
+
+        })
+        
+    }
+
+    //ordenar asc o desc
+    sort():void{
+      this.asc = !this.asc;
+      this.cargarUsuarios();
+    }
+
+    //avanza pagina
+    rewind():void{
+
+      if(!this.isFirst){
+
+        this.page--;
+        this.cargarUsuarios();
+      }
+    }
+
+
+    //retrocede pagina
+    forward():void{
+
+      if(!this.isLast){
+
+        this.page++;
+        this.cargarUsuarios();
+      }
+    }
+
+    //cambia la pagina segun el numero
+    setPage(page:number):void{
+      
+      this.page=page;
+      this.cargarUsuarios();
+    }
+
+    //ordenar por
+    setOrder(order: string): void {
+      this.order = order;
+      this.cargarUsuarios();
+    }
+
     View(usuario:Usuario):void{
-      localStorage.setItem("id",usuario.idusuario.toString());
+      localStorage.setItem("id",usuario.id.toString());
       this.router.navigate(["ViewUsuario"]);
     }
 
     Editar(usuario:Usuario):void{
-      localStorage.setItem("id",usuario.idusuario.toString());
+      localStorage.setItem("id",usuario.id.toString());
       this.router.navigate(["EditUsuario"]);
     }
 
     Delete(usuario:Usuario):void{
-      this.service.deleteUser(usuario).subscribe(data=>{this.usuario=this.usuario?.filter(u=>u!==usuario);
-        alert("Usuario eliminado correctamente"); })
-        this.router.navigate(["getPageUsuario"]);
+
+      swal.fire({
+        title: '¿Estás seguro de que quieres eliminar a ' + usuario.nombreusuario + '?',
+        text: '¡No podrás recuperarla si decides eliminarla!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Sí, quiero eliminarla',
+        cancelButtonText: 'No, quiero guardarla'
+      }).then((result) => {
+        if(result.value){
+
+          this.service.deleteUser(usuario).subscribe(data=>{this.usuario=this.usuario?.filter(u=>u!==usuario);
+            alert("Usuario eliminado correctamente"); })
+         
+
+          swal.fire(
+            '¡Eliminada!',
+            'El usuario ' + usuario.nombreusuario + ' ha sido eliminado.',
+          )
+          } else if (result.dismiss === swal.DismissReason.cancel){
+
+            swal.fire(
+
+              'Cancelado',
+              'El usuario ' + usuario.nombreusuario + ' está segura.',
+            )
+          }
+      })
     }
 
     Volver(){
